pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brain-sidecar"
version = "0.1.0"
description = "All-local live audio transcription, topic notes, and recall sidecar."
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi>=0.115.0",
  "python-multipart>=0.0.9",
  "uvicorn[standard]>=0.30.0",
  "numpy>=1.26.0",
]

[project.optional-dependencies]
gpu = [
  "faster-whisper>=1.1.1",
  "nvidia-cublas-cu12",
  "nvidia-cudnn-cu12==9.*",
]
speaker = [
  "torch>=2.3.0",
  "torchaudio>=2.3.0",
  "speechbrain>=1.0.0",
]
recall = [
  "faiss-cpu>=1.8.0",
  "pypdf>=5.0.0",
  "python-docx>=1.1.0",
]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
brain-sidecar = "brain_sidecar.server.app:main"
brain-sidecar-speaker = "brain_sidecar.tools.speaker_admin:main"

[tool.setuptools.packages.find]
include = ["brain_sidecar*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
